ISSUE #2243 - move check rule to helper
